+++ src/lib/command.ts
@@ -1,1 +1,2 @@
 export type ZdpCommandFieldSize = 'sm' | 'md';
+export type ZdpCommandFieldType = 'search' | 'text';